+++ app/Http/Controllers/BillController.php
@@ -6,20 +6,20 @@ use App\Models\Bill;
 use Inertia\Inertia;
 use App\Http\Requests\BillRequest;
 use App\Http\Responses\BillResponse;
-use App\Actions\Billing\CalculateBill;
-use App\Actions\Billing\ParseBillDetails;
+use App\Contracts\Actions\ParsesBillInformation;
+use App\Contracts\Actions\CalculatesBillInfromation;
 
 class BillController extends Controller
 {
     /**
      * Display the specified resource.
      *
-     * @param \App\Models\Bill                      $bill
-     * @param \App\Actions\Billing\ParseBillDetails $parser
+     * @param \App\Models\Bill                             $bill
+     * @param \App\Contracts\Actions\ParsesBillInformation $parser
      *
      * @return \Illuminate\Http\Response
      */
-    public function show(Bill $bill, ParseBillDetails $parser)
+    public function show(Bill $bill, ParsesBillInformation $parser)
     {
         return Inertia::render('Bills/Show', [
             'details' => $parser->parse($bill),
@@ -29,12 +29,12 @@ class BillController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param \App\Http\Requests\BillRequest     $request
-     * @param \App\Actions\Billing\CalculateBill $calculator
+     * @param \App\Http\Requests\BillRequest                   $request
+     * @param \App\Contracts\Actions\CalculatesBillInfromation $calculator
      *
      * @return \Illuminate\Http\Response
      */
-    public function store(BillRequest $request, CalculateBill $calculator)
+    public function store(BillRequest $request, CalculatesBillInfromation $calculator)
     {
         $data = $request->has('billfile')
             ? file_get_contents($request->billfile)